requirements.txt
torch==2.9.1+cu130
ctransformers==0.2.27
fastapi==0.128.0
huggingface_hub==0.36.0
imageio==2.37.2
moviepy==2.2.1
numpy==2.4.1
Pillow==12.1.0
psutil==7.1.3
pydantic==2.12.5
Requests==2.32.5
streamlit==1.52.2
transformers==4.57.3
uvicorn==0.40.0
